+++ packages/angular-demo/src/app/app.module.ts
@@ -4,7 +4,7 @@ import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { SynthComponentsModule, SynthServicesModule } from 'glyph-angular';
+import { GlyphComponentsModule, GlyphServicesModule } from 'glyph-angular';
 import { HomeComponent } from './components/home/home.component';
 import { environment } from '../environments/environment';
 import { fakeBackendProvider } from '../helpers/fake-backend';
@@ -27,8 +27,8 @@ export function createTranslateLoader(http: HttpClient) {
         HttpClientModule,
         BrowserModule,
         AppRoutingModule,
-        SynthComponentsModule,
-        SynthServicesModule.forRoot(environment),
+        GlyphComponentsModule,
+        GlyphServicesModule.forRoot(environment),
         TranslateModule.forRoot({
             loader: {
                 provide: TranslateLoader,