+++ packages/glyph-angular/projects/glyph-angular/src/glyph-directives/proxies.ts
@@ -6,7 +6,7 @@ import { ProxyCmp, proxyOutputs } from './angular-component-lib/utils';
 import { Components } from 'glyph-components';
 
 
-export declare interface SynthButton extends Components.SynthButton {}
+export declare interface GlyphButton extends Components.GlyphButton {}
 @ProxyCmp({
   inputs: ['icon', 'text']
 })
@@ -16,7 +16,7 @@ export declare interface SynthButton extends Components.SynthButton {}
   template: '<ng-content></ng-content>',
   inputs: ['icon', 'text']
 })
-export class SynthButton {
+export class GlyphButton {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
@@ -25,7 +25,7 @@ export class SynthButton {
 }
 
 import { ChipsBarComponent as IChipsBarComponent } from 'glyph-components/dist/types/components/chipsbar/chipsbar';
-export declare interface SynthChipsbar extends Components.SynthChipsbar {}
+export declare interface GlyphChipsbar extends Components.GlyphChipsbar {}
 @ProxyCmp({
   inputs: ['filtersConfig', 'i18n']
 })
@@ -36,7 +36,7 @@ export declare interface SynthChipsbar extends Components.SynthChipsbar {}
   inputs: ['filtersConfig', 'i18n'],
   outputs: ['filterSelect', 'filterClear', 'filterMultiSelect', 'clearAll']
 })
-export class SynthChipsbar {
+export class GlyphChipsbar {
   /** Filter select event */
   filterSelect!: IChipsBarComponent['filterSelect'];
   /** Filter clear event */
@@ -54,7 +54,7 @@ export class SynthChipsbar {
 }
 
 import { FilterComponent as IFilterComponent } from 'glyph-components/dist/types/components/filter/filter';
-export declare interface SynthFilter extends Components.SynthFilter {}
+export declare interface GlyphFilter extends Components.GlyphFilter {}
 @ProxyCmp({
   inputs: ['description', 'haveMultiSelect', 'i18n', 'multiSelect', 'options', 'plural', 'searchPlaceholder', 'selected']
 })
@@ -65,7 +65,7 @@ export declare interface SynthFilter extends Components.SynthFilter {}
   inputs: ['description', 'haveMultiSelect', 'i18n', 'multiSelect', 'options', 'plural', 'searchPlaceholder', 'selected'],
   outputs: ['optionClickEvent', 'clearEvent', 'multiSelectEvent']
 })
-export class SynthFilter {
+export class GlyphFilter {
   /** Option click event */
   optionClickEvent!: IFilterComponent['optionClickEvent'];
   /** Clear selected filters callback */
@@ -81,7 +81,7 @@ export class SynthFilter {
 }
 
 import { ListComponent as IListComponent } from 'glyph-components/dist/types/components/list/list';
-export declare interface SynthList extends Components.SynthList {}
+export declare interface GlyphList extends Components.GlyphList {}
 @ProxyCmp({
   inputs: ['data', 'decimals', 'defaultSortField', 'enableDownload', 'expandable', 'filterFields', 'i18n', 'limit', 'loading', 'update']
 })
@@ -92,7 +92,7 @@ export declare interface SynthList extends Components.SynthList {}
   inputs: ['data', 'decimals', 'defaultSortField', 'enableDownload', 'expandable', 'filterFields', 'i18n', 'limit', 'loading', 'update'],
   outputs: ['expandRow']
 })
-export class SynthList {
+export class GlyphList {
   /** Expand row event */
   expandRow!: IListComponent['expandRow'];
   protected el: HTMLElement;
@@ -104,7 +104,7 @@ export class SynthList {
 }
 
 
-export declare interface SynthListRow extends Components.SynthListRow {}
+export declare interface GlyphListRow extends Components.GlyphListRow {}
 @ProxyCmp({
   inputs: ['expandHandle', 'expandable', 'fields', 'i18n', 'isTotal', 'row']
 })
@@ -114,7 +114,7 @@ export declare interface SynthListRow extends Components.SynthListRow {}
   template: '<ng-content></ng-content>',
   inputs: ['expandHandle', 'expandable', 'fields', 'i18n', 'isTotal', 'row']
 })
-export class SynthListRow {
+export class GlyphListRow {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
@@ -123,7 +123,7 @@ export class SynthListRow {
 }
 
 
-export declare interface SynthNoData extends Components.SynthNoData {}
+export declare interface GlyphNoData extends Components.GlyphNoData {}
 @ProxyCmp({
   inputs: ['bottomText', 'i18n', 'mode', 'text']
 })
@@ -133,7 +133,7 @@ export declare interface SynthNoData extends Components.SynthNoData {}
   template: '<ng-content></ng-content>',
   inputs: ['bottomText', 'i18n', 'mode', 'text']
 })
-export class SynthNoData {
+export class GlyphNoData {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
@@ -142,7 +142,7 @@ export class SynthNoData {
 }
 
 
-export declare interface SynthSkLoader extends Components.SynthSkLoader {}
+export declare interface GlyphSkLoader extends Components.GlyphSkLoader {}
 @ProxyCmp({
   inputs: ['height', 'repetitions']
 })
@@ -152,7 +152,7 @@ export declare interface SynthSkLoader extends Components.SynthSkLoader {}
   template: '<ng-content></ng-content>',
   inputs: ['height', 'repetitions']
 })
-export class SynthSkLoader {
+export class GlyphSkLoader {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
@@ -161,7 +161,7 @@ export class SynthSkLoader {
 }
 
 
-export declare interface SynthTitle extends Components.SynthTitle {}
+export declare interface GlyphTitle extends Components.GlyphTitle {}
 @ProxyCmp({
   inputs: ['titleText']
 })
@@ -171,7 +171,7 @@ export declare interface SynthTitle extends Components.SynthTitle {}
   template: '<ng-content></ng-content>',
   inputs: ['titleText']
 })
-export class SynthTitle {
+export class GlyphTitle {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
@@ -180,7 +180,7 @@ export class SynthTitle {
 }
 
 
-export declare interface SynthToggler extends Components.SynthToggler {}
+export declare interface GlyphToggler extends Components.GlyphToggler {}
 @ProxyCmp({
   inputs: ['active', 'callback']
 })
@@ -190,7 +190,7 @@ export declare interface SynthToggler extends Components.SynthToggler {}
   template: '<ng-content></ng-content>',
   inputs: ['active', 'callback']
 })
-export class SynthToggler {
+export class GlyphToggler {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();