+++ packages/glyph-react/src/index.tsx
@@ -1,6 +1,6 @@
 import 'glyph-components/dist/glyph-components/glyph-components.css';
 
-import SynthReactList from './wrappers/list/list';
+import GlyphReactList from './wrappers/list/list';
 
 export * from './setup';
 
@@ -8,4 +8,4 @@ export * from './components';
 
 export * from 'glyph-core';
 
-export { SynthReactList };
+export { GlyphReactList };