+++ packages/react-demo/src/App.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { actions, SynthReactList } from 'glyph-react';
+import { actions, GlyphReactList } from 'glyph-react';
 import { getChildrenList, getListData } from './services/data.service';
 import './App.css';
 import { useDispatch } from 'react-redux';
@@ -43,7 +43,7 @@ function App() {
     return (
         <div className="App">
             <div className="list">
-                <SynthReactList
+                <GlyphReactList
                     loading={false}
                     enableDownload
                     expandable